requirements.txt
Appium-Python-Client~=3.1.1